Fix issue with focus ring being clipped
For some reason, the ring gets clipped, even when using what I think should be the correct settings to not do that. I couldn't figure out why it was happening though, so kinda just hacking around it by adding some padding and then a border to try to motivate the padding a little better.